basic graph working

+++ app/src/main/java/com/sam_chordas/android/stockhawk/models/StockDetailModel.java
@@ -15,7 +15,7 @@ public class StockDetailModel {
     }
 
     @SerializedName("count")
-    public long count;
+    public int count;
 
     @SerializedName("created")
     public String created;
